Add a prefix to all generated lists/rules
Add the ability to add a prefix to all generated lists, macros, and
rules. This makes it easier to load multiple rules files generated from
different PSPs without the generated rules/macros/lists overlapping with
each other.

The actual prefix to add is provided as a parameter to
NewConverter(). It can be a blank string, in which case the PSP name is
used as the prefix. If the name is missing for some reason, use the
string "psp".

Signed-off-by: Mark Stemm <mark.stemm@gmail.com>