Limit syscall activity for psp rules to containers
In the case where a psp does not have a falco-rules-psp-images
annotation that limits the images where the rules apply, the syscall
related rules were being applied to both hosts and containers.

Change this so the events are limited to containers.

This also required a few changes to put the evt.type restriction at the
beginning of the rule condition to avoid warnings about evt.types.

Signed-off-by: Mark Stemm <mark.stemm@gmail.com>