update: remove rajibmitra from OWNERS
After changing the github username from rajibmitra to rmitr - Rajib expressed that is very busy to donate time to this project in the near future. We are going to keep Rajib in the org of course but removing from maintainer duty. Rajib, you'll be welcome once you are ready to come back!

Co-Authored-By: Leonardo Di Donato <leodidonato@gmail.com>
Signed-off-by: Lorenzo Fontana <lo@linux.com>