feat(artifact/follow): add new command "artifact follow"
The new command accepts multiple artifact references. For each reference it starts a new go routine
which tracks the artifact and monitors it for changes such new versions. In case of new artifact version
it downloads and installs it in the local filesystem. The sync period can be customized using the "--every, -e"
flag. The default values is set to 24h. Keep in mind that the command does not exit, but runs in daemon mode
until it is stopped by a termination signal.

Signed-off-by: Aldo Lacuku <aldo@lacuku.eu>