fix(follower): do not make assumptions on data retrieved from version endpoint
The follower retrievs the api versions from falco at start up. The api versions
are used to check if an artifact is compatible with the running instance of falco.
Here we expect falco to sent generic data, and we do not make assumptions on the
data type when unmarshalling the received data.

Signed-off-by: Aldo Lacuku <aldo@lacuku.eu>