new(index-caching): add in-memory cache mechanism for index files
Since index files can be overwritten using ENV variables we needed a
new solution to handle them on a per-command basis. Added an in-memory
caching system that handles the index files during the command execution.
Furthermore, it handles loading and saving the index cache from/to filesystem.
The files that tracks the indexes are saved to ~/.config/falcoctl.

Signed-off-by: Aldo Lacuku <aldo@lacuku.eu>
Co-authored-by: Lorenzo Susini <susinilorenzo1@gmail.com>